code/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
click==8.1.3
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.43.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==10.2.0
pyarrow==14.0.1
PyGithub==1.58.2
Pygments==2.15.0
pymongo==4.6.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
unidiff==0.7.3
urllib3==2.0.2
validators==0.20.0
